Refaktorerade lite och skapade klassen SetupView för att göra det lättare att skapa views
