+++ portcdm/src/main/java/sample/AppLayout.java
@@ -39,13 +39,8 @@ public class AppLayout {
         mainStage = primaryStage;
         mainStage.setTitle("Professional Ship Agent Software 2018");
 
-        GridPane grid = Setup_Grid();
-
-        ScrollPane rootPane = new ScrollPane();
-        rootPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.AS_NEEDED);
-        rootPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.AS_NEEDED);
-        rootPane.setContent(grid);
-
+        SetupView.Setup_View();
+        GridPane grid = SetupView.getGrid();
 
         final Text sceneTitle = new Text("Welcome to the portCDM agent application");
         sceneTitle.setFont(Font.font(26));
@@ -64,7 +59,7 @@ public class AppLayout {
         grid.add(sceneTitle, getSceneTitleColumn(), getSceneTitleRow());
         grid.add(hBoxButton, gethBoxButtonColumn(), gethBoxButtonRow());
 
-        defaultView = new Scene(rootPane, getSceneWidth(), getSceneHeight());
+        defaultView = SetupView.getScene();
 
         mainStage.setScene(defaultView);
         mainStage.setResizable(true);
@@ -74,7 +69,7 @@ public class AppLayout {
                 if(choices.getValue() != null && !choices.getValue().toString().isEmpty()){
                     if (choices.getValue().toString().equals("option1")){
                         if(!view1_isCreated){
-                            view1 = View_1.Create_View();
+                            view1 = View_1.Create_View1();
                             view1_isCreated = true;
                         }
                         mainStage.setScene(view1);
@@ -93,14 +88,7 @@ public class AppLayout {
 
     }
 
-    public static GridPane Setup_Grid (){
-        GridPane grid = new GridPane();
-        grid.setAlignment(Pos.CENTER);
-        grid.setHgap(10);
-        grid.setVgap(10);
-        grid.setPadding(new Insets(25, 25, 25, 25));
-        return grid;
-    }
+
     public static HBox Back_Button(){
         HBox button = new HBox(10);
         Button btn = new Button("Back");
